+++ src/components/trailsList.js
@@ -5,7 +5,7 @@ const TrailsList = ({ trails }) => {
   return (
     <section className='page--trails_list'>
       <ul className='trails_list'>
-        {trails.map((trail) => <li key={trail.uuid}>{trail.title}</li>)}
+        {trails.map((trail) => <li key={trail.uuid}>{trail.name}</li>)}
       </ul>
     </section>
   )