+++ src/utils/transforms/trails.transforms.js
@@ -1,0 +1,7 @@
+export function normalizeTrails (trailsArray) {
+  const res = {}
+  for (const trail of trailsArray) {
+    res[trail.uuid] = trail
+  }
+  return res
+}